+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+## 0.94.0 - 2026-07-10
+
 ### Changed
 
 - Breaking: made geometry element collections immutable and element equality/hash exact; use `#eq?(precision:)` for